refactor: 예약 정보 포함 시간 가져오는 api 패키지 이동 (theme -> reservationTime)
