+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -7,8 +7,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.dto.BookedReservationTimeResponseDto;
 import roomescape.dto.ReservationTimeRequestDto;
 import roomescape.dto.ReservationTimeResponseDto;
 import roomescape.service.ReservationTimeService;
@@ -25,6 +27,12 @@ public class ReservationTimeController {
     @GetMapping("/times")
     public List<ReservationTimeResponseDto> readReservationTimes() {
         return reservationTimeService.getAllReservationTimes();
+    }
+
+    @GetMapping("/themes/{themeId}/times")
+    public List<BookedReservationTimeResponseDto> readBookedReservationTimes(
+        @RequestParam String date, @PathVariable Long themeId) {
+        return reservationTimeService.getAllBookedReservationTimes(date, themeId);
     }
 
     @PostMapping("/times")
