+++ src/main/java/roomescape/controller/ThemeController.java
@@ -7,10 +7,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.BookedReservationTimeResponseDto;
 import roomescape.dto.ThemeRequestDto;
 import roomescape.dto.ThemeResponseDto;
 import roomescape.service.ThemeService;
@@ -27,12 +25,6 @@ public class ThemeController {
     @GetMapping("/themes/ranks")
     public List<ThemeResponseDto> readThemeRanks() {
         return themeService.getAllThemeOfRanks();
-    }
-
-    @GetMapping("/themes/{themeId}/times")
-    public List<BookedReservationTimeResponseDto> readBookedReservationTimes(
-        @RequestParam String date, @PathVariable Long themeId) {
-        return themeService.getAllBookedReservationTimes(date, themeId);
     }
 
     @GetMapping("/themes")
