+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,12 +3,8 @@ package roomescape.service;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.dao.ReservationDao;
-import roomescape.dao.ReservationTimeDao;
 import roomescape.dao.ThemeDao;
-import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.dto.BookedReservationTimeResponseDto;
 import roomescape.dto.ThemeRequestDto;
 import roomescape.dto.ThemeResponseDto;
 
@@ -16,15 +12,9 @@ import roomescape.dto.ThemeResponseDto;
 public class ThemeService {
 
     private final ThemeDao themeDao;
-    private final ReservationTimeDao reservationTimeDao;
-    private final ReservationDao reservationDao;
 
-
-    public ThemeService(ThemeDao themeDao, ReservationTimeDao reservationTimeDao,
-        ReservationDao reservationDao) {
+    public ThemeService(ThemeDao themeDao) {
         this.themeDao = themeDao;
-        this.reservationTimeDao = reservationTimeDao;
-        this.reservationDao = reservationDao;
     }
 
     public List<ThemeResponseDto> getAllThemes() {
@@ -41,31 +31,6 @@ public class ThemeService {
     public void deleteTheme(Long id) {
         validateIsExistThemeBy(id);
         themeDao.deleteTheme(id);
-    }
-
-    public List<BookedReservationTimeResponseDto> getAllBookedReservationTimes(String date,
-        Long themeId) {
-        List<ReservationTime> reservationTimes = reservationTimeDao.findAllReservationTimes();
-
-        return reservationTimes.stream()
-            .map(reservationTime -> createBookedReservationTimeResponseDto(date, themeId,
-                reservationTime))
-            .toList();
-    }
-
-    private BookedReservationTimeResponseDto createBookedReservationTimeResponseDto(
-        String date, Long themeId, ReservationTime reservationTime) {
-        if (isAlreadyBookedTime(date, themeId, reservationTime)) {
-            return BookedReservationTimeResponseDto.from(reservationTime, true);
-        }
-        return BookedReservationTimeResponseDto.from(reservationTime, false);
-    }
-
-    private boolean isAlreadyBookedTime(String date, Long themeId,
-        ReservationTime reservationTime) {
-        int alreadyExistReservationCount = reservationDao.calculateAlreadyExistReservationBy(
-            date, themeId, reservationTime.getId());
-        return alreadyExistReservationCount != 0;
     }
 
     private void validateIsExistThemeBy(Long id) {
